Fixed RoomDB errors and shits
- Changed types of ImageVector -> IconOptions and Color -> ColorOptions
- Added converter for IconOptions and ColorOptions
- Changed DateConverter -> Converter and moved it outside to the root data package
- Changes some Int values (specially primary and foreign keys) to Float
- Added index annotations to foreign keys
- Changed AppDatabase configuration
- Added ChecklistRepositoryTest file

- Upgraded version of kotlin 1.9.0 -> 2.0.20 and composeBom 2023.08.00 -> 2024.04.01
- Added KSP and compose (required in kotlin ver 2.0+) plugins

+++ app/src/main/java/com/example/grocerychecklist/data/dao/ChecklistDAO.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.Flow
 interface ChecklistDAO: BaseDAO<Checklist> {
 
     @Query("SELECT * FROM checklist WHERE id = :checklistId LIMIT 1")
-    suspend fun getChecklistById(checklistId: Int): Checklist
+    suspend fun getChecklistById(checklistId: Long): Checklist
 
     @Query("SELECT * FROM checklist ORDER BY lastOpenedAt")
     fun getAllChecklistsOrderedByLastOpenedAt(): Flow<List<Checklist>>
