+++ app/src/main/java/com/example/grocerychecklist/data/dao/HistoryItemDAO.kt
@@ -12,36 +12,36 @@ import java.util.Locale.Category
 interface HistoryItemDAO {
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertBatch(historyItems: List<HistoryItem>): List<Int>
+    suspend fun insertBatch(historyItems: List<HistoryItem>): List<Long>
 
     @Query("SELECT * FROM historyitem WHERE id = :historyItemId LIMIT 1")
-    suspend fun getHistoryItemById(historyItemId: Int): HistoryItem
+    suspend fun getHistoryItemById(historyItemId: Long): HistoryItem
 
     @Query("SELECT * FROM historyitem WHERE historyId = :historyId")
-    fun getAllHistoryItems(historyId: Int): Flow<List<HistoryItem>>
+    fun getAllHistoryItems(historyId: Long): Flow<List<HistoryItem>>
 
     @Query("SELECT * FROM historyitem WHERE historyId = :historyId ORDER BY name")
-    fun getAllHistoryItemsOrderedByName(historyId: Int): Flow<List<HistoryItem>>
+    fun getAllHistoryItemsOrderedByName(historyId: Long): Flow<List<HistoryItem>>
 
     @Query("SELECT * FROM historyitem WHERE historyId = :historyId ORDER BY price")
-    fun getAllHistoryItemsOrderedByPrice(historyId: Int): Flow<List<HistoryItem>>
+    fun getAllHistoryItemsOrderedByPrice(historyId: Long): Flow<List<HistoryItem>>
 
     @Query("SELECT * FROM historyitem WHERE historyId = :historyId ORDER BY `order`")
-    fun getAllHistoryItemsOrderedByOrder(historyId: Int): Flow<List<HistoryItem>>
+    fun getAllHistoryItemsOrderedByOrder(historyId: Long): Flow<List<HistoryItem>>
 
     @Query("SELECT * FROM historyitem WHERE historyId = :historyId AND name = :qName")
-    fun getAllHistoryItemsByName(historyId: Int, qName: String): Flow<List<HistoryItem>>
+    fun getAllHistoryItemsByName(historyId: Long, qName: String): Flow<List<HistoryItem>>
 
     @Query("SELECT * FROM historyitem WHERE historyId = :historyId AND category = :category")
-    fun getAllHistoryItemsByCategory(historyId: Int, category: Category): Flow<List<HistoryItem>>
+    fun getAllHistoryItemsByCategory(historyId: Long, category: Category): Flow<List<HistoryItem>>
 
     @Query("SELECT COUNT(*) FROM historyitem WHERE historyId = :historyId")
-    suspend fun aggregateTotalHistoryItems(historyId: Int): Int
+    suspend fun aggregateTotalHistoryItems(historyId: Long): Int
 
     @Query("""
         SELECT SUM(price)
         FROM historyitem
         WHERE historyId = :historyId
     """)
-    suspend fun aggregateTotalHistoryItemPrice(historyId: Int): Double?
+    suspend fun aggregateTotalHistoryItemPrice(historyId: Long): Double?
 }
