+++ app/src/main/java/com/example/grocerychecklist/data/model/Item.kt
@@ -1,5 +1,6 @@
 package com.example.grocerychecklist.data.model
 
+import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import java.time.LocalDateTime
@@ -7,7 +8,7 @@ import java.time.LocalDateTime
 @Entity
 data class Item(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val name: String,
     val price: Double,
     val category: String,
